Move detection to own static
